+++ crates/sdmx_json/src/primitives.rs
@@ -1,6 +1,7 @@
 use crate::structure::TimeDataType;
 use serde::{Deserialize, Serialize};
 use serde_json::Value;
+use serde_with::serde_as;
 use std::collections::HashMap;
 
 /// A marker trait for denoting that an object is extendable,
@@ -8,6 +9,15 @@ use std::collections::HashMap;
 /// defined in the SDMX-JSON schema.
 pub trait Extendable {
 	fn other(&self) -> Option<&HashMap<String, Value>>;
+}
+
+/// A marker trait for all top-level message types in the
+/// SDMX-JSON standard.
+pub trait SdmxMessage {
+	type Data;
+	fn meta(&self) -> Option<&Meta>;
+	fn data(&self) -> Option<&Self::Data>;
+	fn errors(&self) -> Option<&Vec<StatusMessage>>;
 }
 
 /// A map between languages and the associated content
@@ -262,11 +272,11 @@ pub type Sender = Party;
 pub type Receiver = Party;
 
 /// Non-standard information and basic technical information
-/// associated with a message, which may have a single receiver
-/// (but no more than one).
+/// associated with a message.
 #[derive(Serialize, Deserialize, Debug, Clone, PartialEq, Eq, Default)]
+#[serde_as]
 #[serde(rename_all = "camelCase")]
-pub struct MetaSingleReceiver {
+pub struct Meta {
 	#[serde(skip_serializing_if = "Option::is_none")]
 	pub schema: Option<String>,
 	pub id: String,
@@ -280,36 +290,9 @@ pub struct MetaSingleReceiver {
 	#[serde(skip_serializing_if = "Option::is_none")]
 	pub names: Option<LocalizedText>,
 	pub sender: Sender,
-	#[serde(skip_serializing_if = "Option::is_none")]
-	pub receiver: Option<Receiver>,
-	pub links: Vec<Link>,
-	#[serde(skip_serializing_if = "Option::is_none")]
-	#[serde(flatten)]
-	pub other: Option<HashMap<String, Value>>,
-}
-
-/// Non-standard information and basic technical information
-/// associated with a message, which may have multiple
-/// receivers.
-#[derive(Serialize, Deserialize, Debug, Clone, PartialEq, Eq, Default)]
-#[serde(rename_all = "camelCase")]
-pub struct MetaManyReceivers {
-	#[serde(skip_serializing_if = "Option::is_none")]
-	pub schema: Option<String>,
-	pub id: String,
-	#[serde(skip_serializing_if = "Option::is_none")]
-	pub test: Option<bool>,
-	pub prepared: String,
-	#[serde(skip_serializing_if = "Option::is_none")]
-	pub content_languages: Option<Vec<String>>,
-	#[serde(skip_serializing_if = "Option::is_none")]
-	pub name: Option<String>,
-	#[serde(skip_serializing_if = "Option::is_none")]
-	pub names: Option<LocalizedText>,
-	pub sender: Sender,
+	#[serde_as(as = "OneOrMany<_, PreferOne>")]
 	#[serde(skip_serializing_if = "Option::is_none")]
 	pub receivers: Option<Vec<Receiver>>,
-	#[serde(skip_serializing_if = "Option::is_none")]
 	pub links: Option<Vec<Link>>,
 	#[serde(skip_serializing_if = "Option::is_none")]
 	#[serde(flatten)]
@@ -369,7 +352,6 @@ impl_extendable!(
 	Contact,
 	StatusMessage,
 	Party,
-	MetaSingleReceiver,
-	MetaManyReceivers,
+	Meta,
 	SentinelValue,
 );